Move MenuIcon into the component so it's easier to wrap with a badge without a context
